Move time series file discovery so CI can actually test it
The unit tests added for find_ts_files and ts_files_overlap were guarded
by skipUnless and never ran: the ADF unit test workflow installs only
PyYAML and pytest, while adf_utils imports numpy, xarray, pandas and
geocat at module level.  CI reported green on ten tests it had skipped.

Verified in a venv holding only pyyaml and pytest, which is what
ADF_unit_tests.yaml builds:

  before:  4 failed, 13 passed, 10 skipped
  after:   33 passed
  (the 4 failures are an artifact of extracting only lib/ for the
   comparison - config_cam_baseline_example.yaml lives at the repo root)

The three helpers move to lib/adf_file_utils.py, which imports nothing
but pathlib.  adf_utils re-exports them, so utils.find_ts_files(...) is
unchanged for adf_dataset, adf_derive, adf_info and amwg_table.  No logic
changes - a move plus a re-export.
